+++ app/src/main/java/com/joinz/taskmanager/OnTaskClickListener.java
@@ -1,0 +1,5 @@
+package com.joinz.taskmanager;
+
+public interface OnTaskClickListener {
+    void onClick(Task task);
+}
