+++ app/src/main/res/layout/activity_main.xml
@@ -12,4 +12,11 @@
         tools:listitem="@layout/task_item"
         />
 
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/fabAddTask"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="end|bottom"
+        android:layout_margin="16dp" />
+
 </FrameLayout>
